Fix handling of retractions
When an entity is retracted (or its membership attribute is retracted), it no
longer has that membership attribute in the current db. However it had the
membership attribute in the previos db, which is how it ended up in the table,
so we need to check for membership in both `db` and `prev-db`.

We also had an inconsistency in how we were represention deletions in `:ops`,
using either `:db/id` or `"db__id"`. We now use `"db__id"` consistently, same as
in :upsert.

As an optimization also check for (seq cardinality-{one,many}-datoms).