Freely available deps.
Current deps require being a Datomic customer to get the peer library
and the specified version of datomic-free.

Use the now freely available `com.datomic/peer` dependency for
pro. This is untested as I don't have a Datomic pro system right now.

Also move to the last version of `com.datomic/datomic-free` available
in Maven central. I've verified that the tests pass with this version.